RShiny app created

+++ tensorflow_exps/build.py
@@ -5,6 +5,7 @@ import datetime
 from pathlib import Path
 import shutil
 import os
+import sys
 from tensorflow_exps import log_utils
 
 
@@ -29,7 +30,7 @@ class TF(object):
         # self.data_stats()
         self.model = None
         self.base_stats = ''
-        self.log_dir = ''
+        self.log_dir = self.file_info['save_loc']
 
     def init(self):
         self.read_data()
@@ -63,7 +64,7 @@ class TF(object):
         if 'metrics' not in self.params['tf_compile'].keys():
             self.params['tf_compile']['metrics'] = params_default['tf_compile']['metrics']
         if 'epochs' not in self.params['tf_compile'].keys():
-            self.params['tf_compile']['epochs'] = params_default['epochs']['metrics']
+            self.params['tf_compile']['epochs'] = params_default['tf_compile']['epochs']
 
         if 'create_unique' in self.params.keys():
             assert self.params['create_unique'] in [True, False]
@@ -149,7 +150,7 @@ class TF(object):
             fh.write('\nloss, accuracy metrics: \n')
             for i in fit_history.epoch:
                 fh.write('Epoch {0}/{1} - loss: {2:=0.4f} - accuracy: {3:=0.4f} \n'.format(
-                    i, len(fit_history.epoch), fit_history.history['loss'][i], fit_history.history['accuracy'][i]))
+                    i+1, len(fit_history.epoch), fit_history.history['loss'][i], fit_history.history['accuracy'][i]))
 
             fh.write('\n\n\n' + '#' * 20 + '\n' + 'model summary: \n' + '#' * 20 + '\n')
             self.model.summary(print_fn=lambda x: fh.write(x + '\n'), line_length=80)
@@ -186,16 +187,24 @@ if __name__ == '__main__':
     tf_model = TF(file_info=file_info, params=params)
     tf_model.init()
 
-    # creating the log dir
-    tf_model.log_dir = tf_model.file_info['save_loc'] + \
-                       "logs/fit/{0}_basic_dense_{1}epochs".format(tf_model.params['training_data'],
-                                                                   tf_model.params['tf_compile']['epochs']) + \
-                       tf_model.params['log_dir_suffix']
-    if Path(tf_model.log_dir).exists() and Path(tf_model.log_dir).is_dir():
-        print('log directory already exists, deleting {0}'.format(tf_model.log_dir))
-        shutil.rmtree(tf_model.log_dir)
-    print('creating log directory {0}'.format(tf_model.log_dir))
-    os.makedirs(tf_model.log_dir)
+    # creating the log dir in-situ ---> deprecated
+    diy_log_dir = 0
+    if diy_log_dir:
+        tf_model.log_dir = tf_model.file_info['save_loc'] + \
+                           "logs/fit/{0}_basic_dense_{1}epochs".format(tf_model.params['training_data'],
+                                                                       tf_model.params['tf_compile']['epochs']) + \
+                           tf_model.params['log_dir_suffix']
+        if Path(tf_model.log_dir).exists() and Path(tf_model.log_dir).is_dir():
+            print('log directory already exists, deleting {0}'.format(tf_model.log_dir))
+            shutil.rmtree(tf_model.log_dir)
+        print('creating log directory {0}'.format(tf_model.log_dir))
+        os.makedirs(tf_model.log_dir)
+    elif not(Path(tf_model.log_dir).exists() and Path(tf_model.log_dir).is_dir()):
+        print('creating log directory {0}'.format(tf_model.log_dir))
+        os.makedirs(tf_model.log_dir)
+    else:
+        print('directory already exists, aborting run')
+        sys.exit(0)
 
     tf_model.plot_samples()
     tf_model.build_model()
